refactor video reader backends into their own subpackage
