+++ docs/modules/lightning_pose.data.rst
@@ -9,7 +9,7 @@ lightning\_pose.data
 .. automodapi:: lightning_pose.data.cameras
    :no-inheritance-diagram:
 
-.. automodapi:: lightning_pose.data.dali
+.. automodapi:: lightning_pose.data.video.dali
    :no-inheritance-diagram:
 
 .. automodapi:: lightning_pose.data.datamodules
