+++ lightning_pose/data/video/__init__.py
@@ -1,0 +1,6 @@
+"""Video-reading backends for training and prediction.
+
+TODO(Part 2): package map + "adding a new video reader" contract.
+"""
+
+__all__: list[str] = []
