+++ tests/utils/test_predictions.py
@@ -119,7 +119,7 @@ class TestPredictVideoBboxFile:
 
         with (
             patch('lightning_pose.utils.predictions.count_frames', return_value=3),
-            patch('lightning_pose.data.dali.PrepareDALI', mock_dali),
+            patch('lightning_pose.data.video.dali.PrepareDALI', mock_dali),
             patch('lightning_pose.utils.predictions.pl.Trainer'),
             patch('lightning_pose.utils.predictions.PredictionHandler'),
         ):
@@ -140,7 +140,7 @@ class TestPredictVideoBboxFile:
         mock_dali = MagicMock()
 
         with (
-            patch('lightning_pose.data.dali.PrepareDALI', mock_dali),
+            patch('lightning_pose.data.video.dali.PrepareDALI', mock_dali),
             patch('lightning_pose.utils.predictions.pl.Trainer'),
             patch('lightning_pose.utils.predictions.PredictionHandler'),
         ):
